Initial commit
First try at documentation and two base models.
Not tested for now.

Use rst for README instead of markdown

Remove idea configuration from git

Simplify structure of the app

[+] Added a license file (BSD-3-Clause)
[m] Temporarily changed the models package to mixins
[m] Renamed rectangle mixins to dimension
[+] Added django to requirements

Add pendulum to requirements

[+] Added pendulum (datetime manipulation) to setup.py

Add a requirements file

[f] Fixes ReadTheDocs generation of documentation. On site, virtualenv
was activated for the project and the new requirements file was added
and specified to the RTD admin page.

Put the basemix module in python path

[f] Made the basemix module accessible in the python path instead
of mixins only.

Make basemix the installed package

Add link to documentation